feat: added fee calculation and moved payload computing outside transaction methods
